+++ tidy-resize-images.php
@@ -35,6 +35,7 @@ require_once TRI_PLUGIN_DIR . 'constants.php';
 require_once TRI_PLUGIN_DIR . 'functions-private.php';
 require_once TRI_PLUGIN_DIR . 'includes/class-plugin.php';
 require_once TRI_PLUGIN_DIR . 'includes/class-admin-hooks.php';
+require_once TRI_PLUGIN_DIR . 'includes/class-settings.php';
 require_once TRI_PLUGIN_DIR . 'includes/class-capabilities.php';
 require_once TRI_PLUGIN_DIR . 'includes/class-image-library.php';
 require_once TRI_PLUGIN_DIR . 'includes/class-image-processor.php';